Fix restore app in pause state, power economy mode, add some colors, refactoring

+++ wear/src/main/java/ru/pavel/tomato/wear/timer/TimerState.kt
@@ -4,6 +4,6 @@ import android.os.CountDownTimer
 
 sealed class TimerState {
     data class Active(val timer: CountDownTimer) : TimerState()
-    data class Paused(val timeLeft: Int): TimerState()
+    data class Paused(val timeLeft: Long): TimerState()
     object Idle : TimerState()
 }
